Add the cli::output_option CLI option class
The output_option class adds a new CLI option type to represent the --output
option of the report command.  This option holds in a single place all the
code needed to parse and interpret the descriptor of the output type.

This has no visible changes yet: the code can still only understand
console:/dev/stdout, but this is a preliminary step to adding support for
additional types and locations.

This change was r307 in Subversion.