Generalize SQLite statement parameter binding in a single bind() method
Get rid of all the bind_* methods of sqlite::statement that hardcoded the
type in the function name and use a single overloaded bind() method to
bind any argument type.

Among other things, this introduces a new blob() class to represent the
data related to a BLOB (pointer + size) and a null() class to be able to
bind nulls.  The former will be handy to deal with the storage of test
case logs in the database.

Also, this simplifies a lot of code by removing the need to redefine the
wrapping functions depending on the type.

Need to do the same for column queries, but will happen separately.

This change was r310 in Subversion.