Concatenate test case cleanup output to the regular body output
When running ATF test cases, send the output of the cleanup part to the
same files as the body so that the user can see all the output in a simple
manner.

The current visible effect of this is that the 'debug' command will show
the output of both parts.  Soon, though, we want to store this information
in the database, and it's much easier (and consistent among test interface
implementations) if we only have to store a single file for all parts.

This change was r311 in Subversion.