Add a mechanism to access the stdout and stderr of test cases
When test cases are executed, their stdout and stderr are directed to
temporary files within the work directory of the test case.  These files
are deleted immediately after the test case execution routine finishes,
because the work directory is destroyed as part of the cleanup.

In order to provide access these files, add a "test_case_hooks" class
that allows the caller to hook into the execution process of the test
case.  These hooks can be used from the caller to get a notification
when the stdout and stderr files are ready for processing.

At the moment, the hooks are only used for testing purposes.  The real
goal, though is to implement the storage of the stdout and stderr of
test cases into the database, which will happen separately.

The alternative implementation would have been to explicitly determine
the paths to stdout and stderr when calling run() (just as we do for
debug()).  However, doing so would mean that the caller has to take care
of cleaning up these files (with all the implications this has regarding
signal handling).  In other words, we'd be duplicating the cleanup
functionality of the test cases in order to handle two special cases.
Additionally, in the future, we may also want a way to read arbitrary
files from the test case directory before it is destroyed, so this just
moves into that direction.

This change was r312 in Subversion.