Store the stdout and stderr of test cases into the database
Make 'kyua test' record the output of the test cases (both their stdout and
their stderr) into the database for further inspection.

This is quite rudimentary: I suspect that, because we don't perform any
file deduplication, the database will quickly grow to unmanageable sizes.
But we can evaluate this later once we encounter the problem, as solving
this might be non-trivial or depend an auxiliary library (to generate file
hashes, that is).

As a side-effect of this change: introduce the utils::stream module to
place some shared code between a previously-existing module
(store::backend) and the new code in store::transaction to support the
processing of files.

This change was r315 in Subversion.