Drop second years from copyright notices
File-specific copyright notices are only required to include the year in
which the file was introduced.  There is no need to edit them when the
file changes, which is a relief because keeping them up to date is messy
and fragile.

Therefore, convert all "2010, 2011" to "2010".

This change was r316 in Subversion.