Fix validation of subprocess termination
The tests that ensure that the subprocesses of a children have been killed
by our control algorithms were faulty due to a race: if the test runs so
fast that the system has not yet had a chance to clean up the subprocesses
from the process table, we get a false test failure.  This was very easy
to reproduce under Ubuntu 10.04.1 LTS but not other systems.

To correct this, do as we already were doing in children_test: attempt the
validation a few times with a delay in between to give the system a chance
to clean up the dead processes.  Of course, this still could fail, but the
chances are much lower.

This change was r320 in Subversion.