Implement integer modifiers in the format module.
Add support to specify integer/float filling, width and precision modifiers
in format strings.  This is a bit rudimentary but should cover the small
amount of use cases needed to properly format byte and time quantities
(coming soon).

While here, fix the processing of '%%': avoid stripping such sequences from
the non-format argument (think "foo %s %s" % "%%" % "bar"), and properly
deal with consecutive sequences of '%%'.  The previous feature addition
highlighted the bugs because of the code refactoring, so fixing at the same
time.

This change was r325 in Subversion.