Enable orphaned tests
A recent change to atf-c++ (to be part of 0.16) makes the compiler able to
detect test cases that have been defined but never added to the test suite.
Not surprisingly, this change spotted two tests in Kyua that had never been
executed.  Enable these two tests and fix obvious bugs in one of them.

This change was r326 in Subversion.