Restrict utils::format to %s
Remove support for %c, %d and %f from utils::format and make everything use
%s regardless of the expression type.

These format strings provided no value because the library is driven by the
types of the expressions to format, not what the format string says.  Also,
some format strings were inconsistent (using a format specifier that did not
match the type), so this removes the inconsistency completely.

This change was r327 in Subversion.