Replace some obvious 1s pauses to just 0.1s
These should allow the 3 affected tests to complete much faster in the
majority of the cases, avoiding annoying run stalls that we cannot avoid
until we have parallel execution of tests.

While doing this, increase the number of attempts performed on each wait
so that, on the worst case, we wait for the same amount of time.

Note that the test suite still has some other annoying waits that we
can't easily get rid of yet.  Just addressing the trivial case here.

This change was r328 in Subversion.