Add the utils::units module
The purpose of this new module is to provide mechanisms to format specific
quantities for user consumption and to parse user-provided strings into the
corresponding quantities.

At the moment, the only unit implemented is "bytes", which supports parsing
and formatting strings of the form "123.45K".

Update issue 28
The main purpose of implementing utils::units::bytes is to support the
processing of the require.memory test-case property and to provide a way
to format user-friendly messages when this requirement cannot be fulfilled.

This change was r329 in Subversion.