Add the utils::memory module
This little module provides a function to query the amount of available
physical memory in the system.  This is an independent module instead of
just an adhoc piece of code in the test case's requirements check because
doing this in a portable manner is incredibly ugly.

At the moment, the physical_memory() query has only been implemented for
NetBSD and OS X.  Also, the code compiles fine under Linux.

Update issue 28
The ability to query the amount of available physical memory is required
to implement support for the 'require.memory' property.

This change was r330 in Subversion.