Implement support for the require.memory property
ATF 0.15 introduced a new test case property called require.memory, which
can be used to specify the minimum amount of physical memory needed for
the test case to run.

The need for this condition arose in a NetBSD test that did not make any
sense for machines with few amounts of memory.

Phew.  Adding a new property requires touching way too many places...

Fixes issue 28

This change was r331 in Subversion.