Record timing information of the execution of the test cases
For every test case executed, log its start and end times.  Use this
information to annotate every test case with its duration as part of
the output of 'test'.  Also use this information to calculate the
duration of the failed tests in 'report' and to display the total run
time of the test suite as part of the report.

This change was r334 in Subversion.