Store the absolute path to the test programs
According to the design document of the database (see DatabaseDesign),
test programs are identified by their absolute paths.  Having the
absolute paths aside from the root + relative path fields is important
because different root + relative path combinations can map to the
same test program.  We need a mechanism to quickly locate all entries
mapping to a single test program to collect, e.g. the history of the
test cases of such program.

This change ensures that the absolute paths are recorded in the
database but there are no code changes to make use of this information.
At the moment, this is only to prevent a necessary schema change in
the future.

This change was r335 in Subversion.