Make the report integration tests work without installed files
Allow the integration tests for the report command to run before make
install is executed.  In order for this to work, we must keep all the
KYUA_* environment variables set in the Makefile and not discard them
when invoking 'kyua' within the tests.  But this also means that we
need to account for these variables in the generated reports to
validate their output.

This change was r340 in Subversion.