Allow the store tests to run before installing the schema.sql file
Introduce a new KYUA_STOREDIR environment variable that overrides the
built-in directory from which the schema.sql file is loaded.  This is
mostly useful for running the tests before "make install" is run, as
the Makefile.am will now pass a path to the uninstalled schema.sql file
during "make check".

Along with the previous change, this fixes 'make distcheck' (at least
on NetBSD).

This change was r341 in Subversion.