Ignore environment variables prefixed with _ during testing
The _ environment variable exposes a different behavior when running the
tests in NetBSD and Fedora.  In the case of Fedora, this causes test
breakage because the expected value of _ is different between the test
run and the validation stages.  To fix this, just strip "private"
environment variables from our test and hope things work.

This change was r342 in Subversion.