Initialize a variable to prevent garbage leaking in
When running the physical_memory test on OS X, the resulting value of the
memory is much larger than it should be, later breaking the other tests
that depend on this value.  The cause, unless I'm mistaken, is that the
size_t variable we use for the result is wider than the value really
returned by sysctlbyname.  Initialize such variable to 0 to kill any
garbage that might be left in it.

This change was r344 in Subversion.