Add a --with-atf flag to tweak the build of tests
Building the tests shipped with Kyua is a pretty heavy-weight process
and a high entry barrier due to the difficulty of installing ATF
itself.

To mitigate these problems, introduce support for the --with-atf flag so
that developers and packagers can select whether they want to build the
tests or not.  Having this flag available will also simplify the process
of creating initial Kyua binary packages for certain Linux distributions
(by avoiding the fight of where the test programs should go into).