Include the module's own header file first
This little trick allows us to validate that the header file is
self-contained because no other includes have happened yet.

This change is a followup on e4a3ea698f553849415e57edc899a88f2a267a48
which did the same thing but only for the test programs.  This now
affects the modules themselves.