Add the utils::text module
The new utils::text module provides utility functions to manipulate
strings.  At the moment, it includes a split() function and a to_type()
function.

Use this new module to kill multiple instances of duplicated code across
the project.  Along the way, this should fix a bunch of parsing errors
because the different code duplicates were not handling error conditions
in the exact same manner.