Fix file unprotection for links and omit errors
If we are in a system without lchmod(3) and we have a symlink in the
work directory pointing to a file owned by someone else, the fs::cleanup
routine will fail because our file unprotection using chmod(2) as a
fallback will not succeed.

Fix two issues:
- First, we do not want to run chmod(2) on a symlink.  In the case
  above, if we did have the access rights, we would end up changing the
  permissions of a file pointed to by chmod(2) outside of the work
  directory.  This is obviously a no-no.
- Second, unprotecting files should be a best-effort operation.  Ignore
  any errors from these.  If the unprotection fails for whatever reason,
  we may still succeed with the removal later.

Problem found by Juan RP in kyua-discuss@.  The problem is easy to spot
if Kyua is installed as root on a Linux system and the checks are later
run as non-root.