Allow inline evaluating variables and vectors
Extend the template processor to allow instantiating the value of
variables and vector positions inline with the text.  Otherwise,
having to place these as individual lines makes the raw templates
extremely unreadable (even though if that'd be OK for HTML files
due to whitespace compaction).

Expression evaluation happens by text surrounded by %%.  We allow
any of %%variable%%, %%vector(index)%% and %%length(vector)%%.
These are not valid as statements any more.