Initial addition of rudimentary HTML reports
This change adds a new 'report-html' command to generate static HTML
reports.  This is not done as a modifier of the already-existing
'report' command as planned because the other flags of this command do
not make much sense for HTML reports.

Note that this functionality is incredibly rudimentary at the moment.
The generated reports work though, and it will be nicer to build on
them iteratively.  Also be aware that testing of the HTML output is
very incomplete; at this point, I'm not sure how to achieve this
nicely.