Make 'report' support console-based reports only
The previous incarnations of the 'report' command provided support in
the --output flag to select the format of the output.  However, with
the advent of HTML reports, I find that the behavior of the 'report'
command is actually very different depending on the desired output.

As a result, just make 'report' be the mechanism to generate plain-text
reports that are not supposed to be machine parseable.  We will add
other kinds of reports by providing alternate commands.