Classify kyua program subcommands in categories
Make the output of the 'help' subcommand group information about
subcommands by categories, and adjust the documentation accordingly.

Some commands require a workspace (e.g. a Kyuafile) to operate, while
others operate purely on the database for reporting purposes.  Group
the subcommands this way, which is what the user will be most likely
interested in knowing.