Make utils::text::refill() return a vector
Change the just-introduced refill() function to return a vector of
lines rather than a string with embedded newlines so that the caller
can perform additional transformations on the result (like adding a
tag to the lines).  Preserve the older version as refill_as_string()
which will come in handy in some situations.

While doing this, also add the utils::text::join() auxiliary function
to concatenate a collection of strings into a single string.