Fix corner case in refill
The refill() mechanism broke when the algorithm could not find a place
to split the string and resorted to just copy the rest to the output.
There were two code paths willing to handle this, and one of them was
rightfully wrong and triggered assertion failures.

Fix this by removing the special case of the "last chunk" of the input
from the refill() algorithm, which in turn results in clearer code.