Implement formatting of tabular data
The new utils::text::table module provides a mechanism to describe
tabular data at a high level (by means of the "table" class) and a
way to format the table as plain text for printing.

The main feature of the table formatting is the caller can specify
a maximum width for the table, and the formatting algorithm will
refill a predetermined column of the table to fulfill such width.

This functionality will be used to pretty-print the help messages
of Kyua, by properly aligning options and commands definitions.