Convert text::format_table() into a class
Introduce the text::table_formatter class to replace the standalone
text::format_table() function.  This is to simplify passing around the
definition of the formatting of a table, which will come in very handy
when printing tables to the console.

This new table_formatter class implements a builder patter to allow
setting all the parameters in a single call.