Allow specifying the width of any table column
Change the interface of the table class: instead of providing the
set_max_width and set_refill_column methods to control how the table
is rendered, provide the new set_table_width and set_column_width
methods.

These interfaces are more generic and will allow specifying the width
of any table column, not just the one that has to be refilled.  Note
that this has not been implemented in this change.