Allow setting widths for any table column
This implements the missing functionality in
table_formatter::set_column_width to allow the caller to explicitly set
the width of any table column.  All widths must be equal or greater to
the size of the table cells, except for a single column that can be set
to be automatically refilled.