Tabulate the output of --help
Properly line up the description of the options and commands printed
by --help to make the output more readable.

While doing this, replace the argument text of the --variable flag
from name=value to K=V, as otherwise the identifier for the flag is
too long and makes the output harder to read than it should (by
leaving very few space for the descriptions).