Automatically wrap messages on screen boundaries
Changes the ui() interface to refill messages that do not fit on the
screen.  While doing this, add new out_raw and err_raw methods to
bypass this automatic refilling, and also add out_tag and err_tag
methods to print messages that have a tag to be repeated on multiple
lines.

I am not completely sure this is a good idea.  For some messages this
looks nice (e.g. the output of help) but, for others, it may be more
of a hassle.  Also, this is not enabled for the output of 'report'
just yet.  Maybe should make this easily customizable by the user.