Remove internal constructors for exceptions
The original reason for having the internal constructors for the
exception types (the constructors without parameters) was that the
tree algorithms didn't have access to the full key of the node being
accessed to stick its name into the exception message.

However, because we actually pass the full node key across our
algorithms, we can "flatten" it again to stick the full key into the
exceptions.  Therefore, get rid of the internal constructors and just
create the exceptions with the right messages from the beginning.