Alloc nodes to override how they interact with Lua
Add the push_lua and set_lua methods to the tree nodes so that the
nodes themselves encode the logic to push and set their values from
a Lua state.  This is necessary so that custom user types can actually
work when they do not have a direct correspondence to one of the
predefined node types.