Migrate Kyua's configuration to utils::config
This change puts the utils::config library into use by replacing the
Kyua's custom configuration class (engine::user_files::config) with
the more generic utils::config::tree class.

This diff of this change is huge because of all the call places that
have to be adjusted to deal with a new object type and its new APIs
to query configuration properties.

The semantics of the configuration file remain the same with a little
exception: the properties within the test_suites subtree must now be
strings.  The change is subtle and I don't expect anybody to be
currently using this in their own configuration files, so I have not
bothered to bump the configuration file's version.

The semantics of the --vars flag of the command line have also changed
due to the above: test-suite specific properties now also have to be
prefixed with test_suites.