Pass the whole result to the scan caller
When the scan action operation receives a result, we passed the unpacked
data structure to the caller.  However, this made it very difficult to
pass around some data that we may want to fetch conditionally (e.g. the
output of the test for HTML reports).

Therefore, just return the data structure to the caller and let it decide
what to query.  We can provide the results lazily, which is better for
efficiency.