Implement support for build directories
Add the '--build-root' option to all the commands that deal with the
workspace (debug, list and test).

This option allows the caller to specify where the built test programs
live if the path differs from where the Kyuafiles are located.  This
functionality is specifically designed to be used in conjunction with
the build directories created by a variety of build systems (including
GNU Automake/Autoconf).