Fix handling of doc/paths.texi
The doc/paths.texi file lives in the build directory, not in the source
directory.  Fixes breakage in builds with a build directory.

Also stop distributing this file.  Distributing this file caused "make
distcheck" to not spot the problem, because the references to the file
in the source directory were valid.