Doxyfile should not be distributed
The Doxyfile file should not be part of the distribution.  We ship
Doxyfile.in and Doxyfile is generated by configure.  As part of this
change, fix the path to the Doxyfile: it should really be a path into
the builddir and not the srcdir.

This fixes the build of the project within an object directory with a
clean source tree (not a distfile).  "make distcheck" did not catch the
problem of Doxyfile being searched for in srcdir because we were
distributing Doxyfile.