Use new support for build roots in make check
When invoking 'kyua test' during 'make check', pass the --build-root
option to the tool to point at the top-level directory of the build
tree.  By doing this, we can easily run the tests when builddir is
not the same as srcdir.

This allows us to get rid of the populate-builddir.sh script hack
previously used to copy the Kyuafiles into the build directory in
those cases where the builddir did not match the srcdir.