Initial import of the kyua-atf-compat sources
The kyua-atf-compat module provides tools to ease the transition from
ATF to Kyua.  In particular, this includes a tool to convert Atffile
files to Kyuafile files in an automated manner (atf2kyua), and drop-in
replacement wrappers for both atf-run and atf-report.

Note that the atf2kyua utility is based on the kyuaify.sh script
previously included in the kyua-cli module.