Get rid of the per-test-case run/debug methods
Expose the test case running and debugging functionality through the new
{debug,run}_test_case free functions.

This is a preliminary step to move towards external testers (as described
in TestersDesign), as we need to have a single generic representation for
test cases.  The different implementations will be auto-discoverable
later, at which point the hacks in these free functions will go away.