Get rid of the virtual test case listing method
Change the listing of the test cases in a test program from a virtual
method in the test program to being a set of free functions.

Similarly to the previous change, this is a temporary step towards
external testers.  The current hacks to check for an interface name
will be removed later.