Allow validation of values in leaf nodes
Extend the interface of the config::typed_leaf_node and
config::base_set_node to support redefining a validate() method that
checks a given node value for validity.  This is invoked before updating
the value in the node, and as such provides the mechanism to implement
sanity checks both from set() and set_string().