Introduce a new metadata class
The new metadata class will eventually hold all the metadata of a test
case (or test program!) in a consistent and interface-independent manner.
The idea is that a Kyuafile should be allowed to specify the metadata
for a whole test program (implementing any interface) and get it to work
just as we do with ATF-based test cases.  In general terms, there is
nothing ATF-specific in the set of requirements, for example.

At the moment, this only moves the requirements-related properties.
The rest of the properties will be moved separately.