Get rid of test-case metadata externalization
Now that the metadata is stored at the base test case level, get rid of
the all_properties() method of the test cases and their specialization
and instead just expose the metadata stored in the parent.

This has the side-effect of changing the output of "kyua list -v" to
show property names that do not match what ATF defines, but we will move
in this direction as part of TestersDesign.