Generalize the fake_result functionality of a test
The ATF interface provided a "fake result" functionality.  This is used
whenever the loading of the test case list fails: Kyua generates a fake
test case within the test program to represent this condition, and this
fake test case needs to return a fake result unconditionally.

Due to how TestersDesign will deal with test interfaces, this needs to
be generalized to all test cases.