Add metadata to test programs
Move the timeout definition in the plain-interface test_program class
to the base_test_program class, by means of adding a metadata entry to
the latter.  This should be generalized later to support more metadata
properties, and to let individual tests override them, but for now is
just a step towards getting rid of the test_program class hierarchy.