Generalize the database schema
Stop recording test programs and test cases differently depending on
their interface.  Just assume they all look the same from a data
representation point of view (which, actually, they now do because we
got rid of all the class hierarchies!).

In particular, this introduces a new table to record the metadata of
each test program and test case.

The schema version is hereby bumped to 2.  No backwards-compatibility
code has been implemented to migrate old databases to the new format.
I will consider doing this later, but given the current user base of
Kyua, it's probably not worth the relatively-major effort.