Use test utility functions from ATF
ATF 0.17 has started providing a bunch of utility functions that come
in very handy when writing tests.  Get rid of our own copies of such
functions from utils/test_utils.  The simplification of the Makefiles
is significant.

utils/test_utils cannot be deleted yet due to a single function left
in that file.  Will decide what to do to it in a separate change.