Do not wrap the ouput of non-reporting commands
Wrapping the output of, e.g. the test command, on screen boundaries is
incredibly annoying as (ironically) the output becomes harder to read
and it's impossible to copy/paste a wrapped line properly.

Therefore, disable the wrapping for the db-exec, debug, list and test
commands and leave the screen wrapping on for those commands that print
actual text that needs to be easy to read (like, e.g. help).