Make create_work_directory public
Move the create_work_directory() function from the inner 'detail' namespace
to the engine namespace so that it can be used outside of the isolation
module (as part of the need to call the testers).

I am expecting this to be temporary as either the isolation module needs to
go away, or the way we call the testers generalized.