Use the independent testers to run test cases
This is a preliminary change that modifies the test_case class to use the
independent testers to issue 'run' and 'debug' operations on test cases.
All tests pass at this point, with some minor adjustments to expected
strings.

This change obsoletes lots of code, and such code will be pruned separately.
Many of the tests for the code to-be-removed still need migration to the new
code paths, hence why I'll reevaluate this later and iteratively.