Redistribute tests in atf_iface and plain_iface
Merge the tests in the various *_iface directories into other test programs
and locations so that we can remove the *_iface directories themselves.
This change actually kills the directories (finally!).

The tests for the test case execution have been blindly moved to
test_case_test.  These should be "audited" for duplication and actual
usefulness.  The test_program_* helpers have been moved to the bootstrap
directory, as this was the only place where they were used.