Initial import of the kyua-testers sources
The kyua-testers module provides scriptable interfaces to interact with
test programs of various kinds.  The interface of such testers allows
the caller to execute a single test case of a single test program in a
controlled and homogeneous manner.

I am discarding all previous development history of kyua-testers as part
of this initial import, as the history contained no useful information
nor reasonable changesets.

Update issue 47
This new module provides the initial implementation of the independent
testers originally described in the TestersDesign project three months
ago.