Merge the child_with_{files,output} classes
Merge the two classes to represent child processes, child_with_files and
child_with_output, into a single child base class.  The only difference
is in the fork methods and in the ability to query the output stream.
None of these deserve the complexity of having two classes.

This will simplify the abstraction to execute testers by allowing the
code to share a single base type for all kinds of children.