Remove support to enforce timeouts on subprocesses
Kill the ability to specify a timeout to wait() calls on subprocesses.
This functionality has been moved to the kyua-testers, which we are
going to trust to do the right thing and not get stuck.

There is still one case in the code where this was useful: namely, the
wait for the external invocation of umount(8).  However, this code is
going away soon too, because it has been moved to kyua-testers.

As a side-effect of this change, we can remove the orphaned
utils/signals/timer module.