Remove the isolation module
Change the test case execution code to use the new interrupts handling
code in utils/signals/interrupts and remove the engine/isolation module.

The isolation module was overly-complex with its hook-based execution
and, worse, it had race conditions that meant that sometimes interrupts
would not be caught properly.

As a side-effect of this change, the utils::signals::programmer class
has been changed to program signals with SA_RESTART.  This brings
further simplifications to the code due to the removal of needless EINTR
handling.