Add {mkdtemp,mkstemp} methods to the auto cleaners
Add a new mkdtemp method to auto_directory and a mkstemp method to
auto_file to provide an "atomic API" to create self-destructing
temporary directories and files.

This requires changing the auto_* classes to allow shallow-copying,
so that the new auxiliary function can return an auto_* object.

While doing this, also add a fs::unlink method to abstract the error
handling of file removal.  We need this code path in two places
after this change, and thus this becomes useful.