Move work directory creation to the drivers
Change the run_test_case and debug_test_case functions to not create a
work directory.  Instead, pass this in from the drivers so that we use
the same work directory for all tests.

This reduces the overhead of having to re-create a work directory for
every test.

Change the various temporary files created during the execution of the
tests to be created inside this work directory, and to be automatically
removed.