Simplify fs::cleanup as fs::rm_r
Remove all the complex logic behind fs::cleanup to remove the work
directory of test cases.  This involved coping with possible mount
points and files whose permissions were hardened.  The cleanup of test
case work directories is now performed in the kyua-testers.

Because we still need the ability to recursively remove file hierarchies
(for the report-html command), add a new fs::rm_r function that does
just this in a very simple manner.