Generalize test program parsing in Kyuafiles
Decouple the parsing of test program definitions in Kyuafiles from the
specific interfaces they implement.

To do this, change engine::tester_path() to dynamically look for all
available testers and cache their information.  This is to support
queries of the form "is this interface supported?", which is what we
need for this change.