Optimize reports
Add a missing primary key and two necessary database indexes so that the
querying of all test programs and test cases of a single action is fast.

This makes "kyua report" (as well as "kyua report-html") bearable.  They
had never been fast, but with the conversion to the independent testers
they became unusable.  They are now faster than ever.

With this change, running "kyua report" after a "make check" on Kyua
itself takes now 2 seconds instead of the 2 minutes it used to take.