Increase precision of datetime::delta
Bump the size of the seconds quantity in datetime::delta to be able to hold
longer deltas.  This is to reduce the possible confusion introduced by the
difference in precision of the datetime::timestamp class.

While doing so, rename {from,to}_useconds to {from,to}_microseconds, as this
is easier to read and matches the names used in datetime::timestamp.