Implement schema migration facilities
With the move to independent testers, we introduced incompatible
schema changes (the new metadatas table among several other details)
and bumped the schema version to 2.

This change adds a new migrate_schema function, accompanied by
supporting SQL code and abundant tests, to take a database with at
schema version 1 and migrate it to version 2.

The migration schema facilities are not yet exposed to the user.