Add the db-migrate command
db-migrate provides the frontend to the database schema migration
algorithm, and allows users to upgrade an old database to the current
version of the schema.

This also modifies the error reporting in kyua to spot old versions
of the schema and report them in a nice manner to the user.