Fix some tests in Mac OS X 10.8
Due to /tmp being a symlink in OS X, we cannot perform string
comparisons on expected paths that lie within the temporary directory.
Doing so causes spurious failures because getcwd(3) cannot know about
symlinks.

(We could be fancier by normalizing the offending paths and comparing
the results... but that's too much work for little benefit in a couple
of ad-hoc test cases.)