Rewrite the Kyuafile file parsing in C++
The previous implementation in Lua was extremely confusing due to an
early attempt to generalize the parsing of various configuration files
(e.g. the init.lua module and the management of the syntax() call).

The new implementation in C++ is straightforward: it just exposes a
set of functions to Lua and uses different Lua states to parse every
nested Kyuafile.  There is no need to attempt to clear the Lua
environment in every file inclusion, which was confusing.

This has the side-effect of removing the init and common modules from
engine::user_files, and in turn makes the 'kyua' binary self-contained.

Update issue 37
The main goal of this change is to get rid of the common code to parse
syntax()s calls and, also, the attempt to generalize file parsing
through a single parser.  Note that config files were already not using
this code.

With this done, removing the name of the file format from syntax() calls
becomes possible -- which is what this issue requested.