Remove format names from syntax() calls
Change the prototype of the syntax() calls in both the Kyuafile
and configuration file formats to only take a version number.

The format is implicitly determined by the context in which the file is
being loaded, and thus such parameter was useless.  The worst part is
that the existence of the parameter caused some end-user confusion.

Unfortunately, this cannot be done in a way to allow existing binaries
to complain about a bad version number, because this changes the
specific thing used by configuration files to denote their format.
Therefore, just ride the current version number without bumping it:
new binaries can deal with old-style files fine, but old binaries will
fail when they encounter the new-style files.

Fixes issue 37.