Bump version of config and kyuafile files to 2
Revision a4d7f7fddf8caaad408f5bddbf09149017b7f010 changed the format of
the syntax() calls to only take the version number without the format
name.  That change should have bumped the version number of these files
to 2, if only to clearly denote to users that there has been a change.

Do this version bump now and ensure proper parsing of old-style
definitions and new-style ones.  The kyua tool remains fully compatible
with the old files.

Update issue 37
Please note that, as part of this issue, the versions of the syntax of
the config and kyuafile files has been bumped to 2 for clarity reasons.