Fix core dump gathering in Linux
First of all, look for core files simply named "core" in the directory
where the program was executed, as this is the default naming scheme in
Linux.  (It'd be nice if we parsed /proc/sys/kernel/core_* and used
those to find the exact core file... but this is "very hard".)

Second, provide an empty stdin to gdb (instead of just closing it).
The version I have (7.4.1-debian) crashes otherwise.

And lastly, avoid double-flusing of stdout by ensuring that the
stdout stream is flushed *before* the fork to run gdb.