Cache configure results for custom tests
Add caching of the results of most of our custom tests -- particularly
to those that need to execute a test program to determine the outcome
of the result.

This is to aid cross-build environments to pass the expected result of
these configure tests, as they cannot execute the tests themselves.