Fix schema_inttest:current_schema in distcheck
The schema_inttest:current_schema test was failing during distcheck (and
presumable also during "make installcheck", which I hadn't tried for a
long time) because it was looking for the current schema file
(schema_v2.sql) in the directory of the test case.  This is wrong,
because the current schema lives in the datadir.