Escape backslashes in test metadata
The previous code in kyua-atf-tester escaped single quotes in the
metadata of test cases so that those single quotes did not yield
invalid Lua strings in the output of the tester.

It turns out we also need to escape backslashes for things to work
properly.  Backslashes also have special meaning within Lua strings.

Found while running the NetBSD test suite.  In particular, the
lib/libc/gen/t_fnmatch test program had the '\'' sequence in the
description of a test and this made the test program be reported
as bogus.