Be more resilent to tester failures
If a tester fails or crashes, ensure we report the problem in a controlled
manner rather than exiting kyua abruptly.

In particular, if the tester fails while running a test, mark the test as
broken.  If the tester fails while listing the test cases of a test
program, mark the test program as broken.  (The latter was already being
done but this change ensures that the condition is properly tested for.)